refactor: split the data table view in the example
